+++ crates/miden-tx/src/host/script_mast_forest_store.rs
@@ -12,7 +12,7 @@ use vm_processor::MastForestStore;
 /// transaction and input note scripts.
 pub struct ScriptMastForestStore {
     mast_forests: BTreeMap<Word, Arc<MastForest>>,
-    adv_data: AdviceMap,
+    advice_map: AdviceMap,
 }
 
 impl ScriptMastForestStore {
@@ -23,7 +23,7 @@ impl ScriptMastForestStore {
     ) -> Self {
         let mut mast_store = ScriptMastForestStore {
             mast_forests: BTreeMap::new(),
-            adv_data: AdviceMap::new(),
+            advice_map: AdviceMap::new(),
         };
 
         for note_script in note_scripts {
@@ -45,13 +45,13 @@ impl ScriptMastForestStore {
 
         // collect advice data from the forest
         for (key, values) in mast_forest.advice_map().clone() {
-            self.adv_data.insert((*key).into(), values);
+            self.advice_map.insert((*key).into(), values);
         }
     }
 
     /// Returns a reference to the advice data collected from all forests.
-    pub fn advice_data(&self) -> &AdviceMap {
-        &self.adv_data
+    pub fn advice_map(&self) -> &AdviceMap {
+        &self.advice_map
     }
 }